implement singleplayer features
