+++ client/database/index.jsx
@@ -9,7 +9,7 @@ import DifficultyDropdown from '../scripts/components/DifficultyDropdown.min.js'
 import Star from '../scripts/components/Star.min.js';
 import CategoryManager from '../scripts/utilities/category-manager.js';
 import { getDropdownValues } from '../scripts/utilities/dropdown-checklist.js';
-import { insertTokensIntoHTML } from '../scripts/utilities/insert-tokens-into-html.js';
+import { insertTokensIntoHTML } from '../insert-tokens-into-html.js';
 
 const starredTossupIds = new Set(await star.getStarredTossupIds());
 const starredBonusIds = new Set(await star.getStarredBonusIds());
